feat: auto discover a config
